Increase default AMSDU buffer size iwlwifi
From Josh Schmelzle:
default buffer for ax210 on ucode 66/67 is amsdu_size:amsdu size 0: 12K
for multi Rx queue devices, 2K for AX210 devices, 4K for other devices
1:4K 2:8K 3:12K (16K buffers) 4: 2K (default 0) (int).

so, let's set it to 3 for a 12K buffer. based on my testing, this is a
huge improvement when capturing 802.11 data frames. the thought here is
the wrong default buffer size was chosen for the ax210.

Signed-off-by: Daniel Finimundi <daniel@finimundi.com>